Refactor ShopStack detail helpers into modular layer
