+++ src/ui/views/browser/components/shopstack/detail/index.js
@@ -1,0 +1,9 @@
+export * from './formatting.js';
+export * from './effects.js';
+export * from './requirements.js';
+export * from './requirementList.js';
+export * from './highlights.js';
+export * from './sections.js';
+export * from './statusHeader.js';
+export * from './cta.js';
+export * from './view.js';
